+++ src/torch_attention/__init__.py
@@ -1,6 +1,10 @@
 from importlib.metadata import PackageNotFoundError, version
 
+from torch_attention.shared import ScaledDotProductAttention
+
+__all__ = ["ScaledDotProductAttention", "__version__"]
+
 try:
-    __version__ = version(__name__)
+    __version__ = version("torch-attention")
 except PackageNotFoundError:
     __version__ = "0.0.0"